backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
python-dotenv
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
deepface
tf-keras
google-generativeai
faiss-cpu
pypdf2
numpy
requests
